updating publish service
