--- ebookapp/src/main/java/com/abk/ebook/ebookapp/EbookappApplication.java
@@ -1,17 +1,0 @@
-package com.abk.ebook.ebookapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
-
-@SpringBootApplication
-public class EbookappApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EbookappApplication.class, args);
-	}
-	
-	
-
-}
