--- ebookapp/src/main/java/com/abk/ebook/ebookapp/exception/NotFoundException.java
@@ -1,13 +1,0 @@
-package com.abk.ebook.ebookapp.exception;
-
-public class NotFoundException extends Exception {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	public NotFoundException(String msg) {
-		super(msg);
-	}
-}
