--- ebookapp/src/main/java/com/abk/ebook/ebookapp/service/PdfGenerationService.java
@@ -1,40 +1,0 @@
-package com.abk.ebook.ebookapp.service;
-
-
-
-import com.itextpdf.text.Document;
-import com.itextpdf.text.DocumentException;
-import com.itextpdf.text.Paragraph;
-import com.itextpdf.text.pdf.PdfWriter;
-import org.springframework.stereotype.Service;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-
-@Service
-public class PdfGenerationService {
-
-    public byte[] generatePdfForBook(byte[] bookContent) {
-        try (ByteArrayOutputStream outputStream = new ByteArrayOutputStream()) {
-            generatePdf(bookContent, outputStream);
-            return outputStream.toByteArray();
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null; // Or handle the exception as needed
-        }
-    }
-
-    private void generatePdf(byte[] bookContent, ByteArrayOutputStream outputStream) throws IOException {
-        Document document = new Document();
-        try {
-            PdfWriter.getInstance(document, outputStream);
-            document.open();
-            document.add(new Paragraph(new String(bookContent)));
-        } catch (DocumentException e) {
-            e.printStackTrace();
-        } finally {
-            document.close();
-        }
-    }
-}
-
