--- ebookapp/src/main/resources/application.properties
@@ -1,19 +1,0 @@
-server.port=8093
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/ebook?createDatabaseIfNotExist=true
-spring.datasource.username=root
-spring.datasource.password=
-
-
-
-
-#Hibernate properties
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql = true
-
-#thyme leaf properties
-spring.thymeleaf.enabled=true
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
-
