--- ebookapp/src/test/java/com/abk/ebook/ebookapp/EbookappApplicationTests.java
@@ -1,13 +1,0 @@
-package com.abk.ebook.ebookapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EbookappApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
