+++ publishservice/src/main/java/com/example/publish/PublishServiceApplication.java
@@ -1,0 +1,13 @@
+package com.example.publish;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PublishServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PublishServiceApplication.class, args);
+	}
+
+}
