+++ publishservice/src/main/java/com/example/publish/exceptions/NotFoundException.java
@@ -1,0 +1,13 @@
+package com.example.publish.exceptions;
+
+public class NotFoundException extends Exception {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+	public NotFoundException(String msg) {
+		super(msg);
+	}
+}
