+++ publishservice/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+
+server.port=8093
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/ebookpublish?createDatabaseIfNotExist=true
+spring.datasource.username=root
+spring.datasource.password=
+
+
+
+#Hibernate properties
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql = true
+
+
