+++ publishservice/src/test/java/com/example/publish/PublishServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.publish;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PublishServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
